requirements.txt
celery==5.3.1
Django==4.0
django-q==1.3.9
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
openapi-codec==1.3.2
openapi-python-client==0.14.1
PyJWT==1.7.1
pytest==7.4.0
redis==3.5.3
requests==2.31.0
